requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
distlib==0.3.8
filelock==3.16.1
idna==3.10
mysql-connector==2.2.9
numpy==2.1.2
pandas==2.2.3
platformdirs==4.3.6
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
